vue project

+++ laravel_api/app/Http/Controllers/Api/V2/TaskController.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
+use Illuminate\Support\Facades\Gate;
 
 // use App\Http\Store\StoreTaskRequest;
 // use App\Http\Requests\UpdateTaskRequest;
@@ -19,6 +20,8 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -27,6 +30,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if(request()->user()->cannot('create', Task::class)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
         
         return TaskResource::make($task);
@@ -37,6 +44,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
+
         return TaskResource::make($task);
     }
 
@@ -45,6 +54,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if($request->user()->cannot('update', $task)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -55,6 +68,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if(request()->user()->cannot('delete', $task)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->delete();
 
         return response()->noContent();
